fix: generic type for deployment and federation
